vtok_p11: Implement internal token update and reload
This commit introduces an internal mechanism for updating and reloading
tokens within the vtok_p11 backend, eliminating the need to reload nginx
for token refreshes or updates. Whenever a session object is accessed
with mutable permissions, the `try_reload` function associated with the
device checks if a new configuration is available. If a new
configuration is detected, it is loaded, and all token information is
updated internally within the p11 token provider.

Signed-off-by: Erdem Meydanli <meydanli@amazon.com>